requirements.txt
# ============================================================
# EcoMobility Assistant — RAG System Dependencies
# Install dengan: pip install -r requirements.txt
# ============================================================

# Core Data Processing
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.10.0  # For Excel file support

# RAG & NLP
langchain>=0.1.0
langchain-community>=0.0.1
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.6.0

# Document Loaders
pypdf>=3.17.0

# LLM Integration
google-genai>=0.3.0

# Web & UI
streamlit>=1.28.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
